Revert "(fix) SpendLogs Table"
This reverts commit 2a1104d1cfef826e18637234e7c2f983d4b06826.

+++ litellm/proxy/_types.py
@@ -346,9 +346,9 @@ class LiteLLM_SpendLogs(LiteLLMBase):
     model: Optional[str] = ""
     call_type: str
     spend: Optional[float] = 0.0
-    total_tokens: Optional[int] = 0
-    prompt_tokens: Optional[int] = 0
-    completion_tokens: Optional[int] = 0
+    total_tokens: Optional[float] = 0.0
+    prompt_tokens: Optional[float] = 0.0
+    completion_tokens: Optional[float] = 0.0
     startTime: Union[str, datetime, None]
     endTime: Union[str, datetime, None]
     user: Optional[str] = ""
